# swagger.yaml
openapi: 3.0.0
info:
  title: DeepThought_NudgeAPI
  contact: {}
  version: '1.0'
servers:
# - url: http://localhost:5000/api/v3/app
- url: https://deepthought-nudge-api.onrender.com/api/v3/app
  variables: {}
paths:
...
  /events:
    get:
      tags:
      - Events
      summary: Get All Events
      operationId: GetAllEvents
      parameters:
      - name: limit
        in: query
        description: ''
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: page
        in: query
        description: ''
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: type
        in: query
        description: ''
        style: form
        explode: true
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    post:
      tags:
      - Events
      summary: Create Event
      operationId: CreateEvent
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateEventRequest'
              - example:
                  category: ''
                  subcategory: ''
                  name: ''
                  user: ''
                  moderator: ''
                  rigor_rank: 5
                  attendees:
                  - ''
                  imageUrl: ''
                  description: ''
                  scheduled:
                  - on: 2024-07-11
                    from: 2024-07-11 9:00:00
                    to: 2024-07-11 13:00:00
                  tagline: Skydiving is fun. Skydiving is enthralling.
            example:
              category: ''
              subcategory: ''
              name: ''
              user: ''
              moderator: ''
              rigor_rank: 5
              attendees:
              - ''
              imageUrl: ''
              description: ''
              scheduled:
              - on: 2024-07-11
                from: 2024-07-11 9:00:00
                to: 2024-07-11 13:00:00
              tagline: Skydiving is fun. Skydiving is enthralling.
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /nudge:
    post:
      tags:
      - Nudge
      summary: Create Nudge
      operationId: CreateNudge
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateNudgeRequest'
              - example:
                  eventCategoryId: ''
                  eventId: ''
                  title: ''
                  icon: ''
                  scheduledOn: ''
                  scheduledFrom: ''
                  scheduledTo: ''
                  coverImage: ''
                  invitationLine: ''
            example:
              eventCategoryId: ''
              eventId: ''
              title: ''
              icon: ''
              scheduledOn: ''
              scheduledFrom: ''
              scheduledTo: ''
              coverImage: ''
              invitationLine: ''
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - Nudge
      summary: Get All Nudges
      operationId: GetAllNudges
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /nudge/{id}:
    put:
      tags:
      - Nudge
      summary: Update Nudge
      operationId: UpdateNudge
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateNudgeRequest'
              - example:
                  eventCategoryId: ''
                  eventId: ''
                  title: ''
                  icon: ''
                  scheduledOn: ''
                  scheduledFrom: ''
                  scheduledTo: ''
                  coverImage: ''
                  invitationLine: ''
            example:
              eventCategoryId: ''
              eventId: ''
              title: ''
              icon: ''
              scheduledOn: ''
              scheduledFrom: ''
              scheduledTo: ''
              coverImage: ''
              invitationLine: ''
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - Nudge
      summary: Delete Nudge
      operationId: DeleteNudge
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - Nudge
      summary: Get Single Nudge By Id
      operationId: GetSingleNudgeById
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    CreateEventRequest:
      title: CreateEventRequest
      required:
      - category
      - subcategory
      - name
      - user
      - moderator
      - rigor_rank
      - attendees
      - imageUrl
      - description
      - scheduled
      - tagline
      type: object
      properties:
        category:
          type: string
        subcategory:
          type: string
        name:
          type: string
        user:
          type: string
        moderator:
          type: string
        rigor_rank:
          type: integer
          format: int32
        attendees:
          type: array
          items:
            type: string
          description: ''
        imageUrl:
          type: string
        description:
          type: string
        scheduled:
          type: array
          items:
            $ref: '#/components/schemas/Scheduled'
          description: ''
        tagline:
          type: string
      example:
        category: ''
        subcategory: ''
        name: ''
        user: ''
        moderator: ''
        rigor_rank: 5
        attendees:
        - ''
        imageUrl: ''
        description: ''
        scheduled:
        - on: 2024-07-11
          from: 2024-07-11 9:00:00
          to: 2024-07-11 13:00:00
        tagline: Skydiving is fun. Skydiving is enthralling.
    Scheduled:
      title: Scheduled
      required:
      - on
      - from
      - to
      type: object
      properties:
        on:
          type: string
        from:
          type: string
        to:
          type: string
      example:
        on: 2024-07-11
        from: 2024-07-11 9:00:00
        to: 2024-07-11 13:00:00
...
    CreateNudgeRequest:
      title: CreateNudgeRequest
      required:
      - eventCategoryId
      - eventId
      - title
      - icon
      - scheduledOn
      - scheduledFrom
      - scheduledTo
      - coverImage
      - invitationLine
      type: object
      properties:
        eventCategoryId:
          type: string
        eventId:
          type: string
        title:
          type: string
        icon:
          type: string
        scheduledOn:
          type: string
        scheduledFrom:
          type: string
        scheduledTo:
          type: string
        coverImage:
          type: string
        invitationLine:
          type: string
      example:
        eventCategoryId: ''
        eventId: ''
        title: ''
        icon: ''
        scheduledOn: ''
        scheduledFrom: ''
        scheduledTo: ''
        coverImage: ''
        invitationLine: ''
    UpdateNudgeRequest:
      title: UpdateNudgeRequest
      required:
      - eventCategoryId
      - eventId
      - title
      - icon
      - scheduledOn
      - scheduledFrom
      - scheduledTo
      - coverImage
      - invitationLine
      type: object
      properties:
        eventCategoryId:
          type: string
        eventId:
          type: string
        title:
          type: string
        icon:
          type: string
        scheduledOn:
          type: string
        scheduledFrom:
          type: string
        scheduledTo:
          type: string
        coverImage:
          type: string
        invitationLine:
          type: string
      example:
        eventCategoryId: ''
        eventId: ''
        title: ''
        icon: ''
        scheduledOn: ''
        scheduledFrom: ''
        scheduledTo: ''
        coverImage: ''
        invitationLine: ''
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...